requirements.txt
gradio==6.9.0
python-dotenv==1.2.2

langchain==1.2.10
langchain-core==1.2.17
langchain-community==0.4.1
langchain-openai==1.1.10
langchain-chroma==1.1.0

chromadb==1.5.5

faiss-cpu==1.13.2
openai==2.25.0

youtube-transcript-api==1.2.4
tiktoken==0.12.0

rank_bm25==0.2.2

sentence-transformers==3.0.1